+++ src/main/java/cz/charwot/chasify/repositories/ITaskRepository.java
@@ -10,8 +10,8 @@ public interface ITaskRepository {
     void create(Task task);
     void update(Task task) throws IllegalArgumentException;
     void delete(Task task);
-    void deleteById(Long id) throws IllegalArgumentException;
-    Task findById(Long id);
+    void deleteById(int id) throws IllegalArgumentException;
+    Task findById(int id);
     List<Task> findAll();
     List<Task> findByProject(Project project);
     List<Task> findByAssignee(User assignee);
